+++ src/App.tsx
@@ -23,7 +23,7 @@ import AdminSessionPage from './pages/admin/AdminSessionPage';
 import CohortDashboard from './pages/admin/CohortDashboard';
 import CreateUpdateCohortPage from './pages/admin/CreateUpdateCohortPage';
 import BulkCohortUploadPage from './pages/admin/BulkCohortUploadPage';
-import CohortStatsDashboard from './components/CohortStatsDashboard';
+import CohortStatsDashboard from './components/cohorts/CohortStatsDashboard';
 
 const App: React.FC = () => {
   return (
